+++ universal-login-react/src/ui/WalletSelector/Suggestions/index.ts
@@ -1,0 +1,1 @@
+export {Suggestions} from './Suggestions';
